Refactor interpolators and rendering methods to use Precision utility for improved accuracy

+++ src/client/java/tytoo/weave/animation/EasingFunction.java
@@ -2,5 +2,5 @@ package tytoo.weave.animation;
 
 @FunctionalInterface
 public interface EasingFunction {
-    float ease(float t);
+    double ease(double t);
 }
